requirements.txt
Django==2.0
djangorestframework==3.7.3
olefile==0.44
Pillow==4.3.0
pytz==2017.3
coverage==4.4.2